Separated Msgs enumeration from Progam.cs
